Move shared types to shared folder
